Moved styleguide html renderer in to a styleguide helper, updated nested panels to better represent a real example
