+++ app/controllers/koi/styleguide_controller.rb
@@ -1,4 +1,5 @@
 module Koi
   class StyleguideController < ApplicationController
+    helper "koi/styleguide"
   end
 end
